refactor: split structured text parsing
